+++ meteor/client/ui/Status/SystemStatus.tsx
@@ -8,7 +8,7 @@ import {
 import * as reacti18next from 'react-i18next'
 import * as i18next from 'i18next'
 import Moment from 'react-moment'
-import { assertNever, getCurrentTime, getHash, unprotectString } from '../../../lib/lib'
+import { assertNever, getCurrentTime, getHash, protectString, unprotectString } from '../../../lib/lib'
 import { Link } from 'react-router-dom'
 import Tooltip from 'rc-tooltip'
 import { faTrash, faEye } from '@fortawesome/free-solid-svg-icons'
@@ -34,6 +34,8 @@ import { isTranslatableMessage, translateMessage } from '@sofie-automation/corel
 import { i18nTranslator } from '../i18n'
 import { SchemaForm } from '../../lib/forms/schemaForm'
 import { CoreSystem, PeripheralDevices } from '../../../lib/clientCollections'
+import { PeripheralDeviceId } from '@sofie-automation/shared-lib/dist/core/model/Ids'
+import { DebugStateTable } from './DebugState'
 
 interface IDeviceItemProps {
 	// key: string,
@@ -41,6 +43,8 @@ interface IDeviceItemProps {
 	showRemoveButtons?: boolean
 	toplevel?: boolean
 	hasChildren?: boolean
+
+	debugState: object | undefined
 }
 interface IDeviceItemState {}
 
@@ -232,6 +236,8 @@ export const DeviceItem = reacti18next.withTranslation()(
 							</div>
 						</div>
 					) : null}
+
+					{this.props.debugState ? <DebugStateTable debugState={this.props.debugState} /> : null}
 
 					<div className="actions-container">
 						<div className="device-item__actions">
@@ -501,7 +507,9 @@ export const CoreItem = reacti18next.withTranslation()(
 interface ISystemStatusProps {}
 interface ISystemStatusState {
 	systemStatus: StatusResponse | undefined
+	deviceDebugState: Map<PeripheralDeviceId, object>
 }
+
 interface ISystemStatusTrackedProps {
 	coreSystem: ICoreSystem | undefined
 	devices: Array<PeripheralDevice>
@@ -523,6 +531,7 @@ export default translateWithTracker<ISystemStatusProps, ISystemStatusState, ISys
 		ISystemStatusState
 	> {
 		private refreshInterval: NodeJS.Timer | undefined = undefined
+		private refreshDebugStatesInterval: NodeJS.Timer | undefined = undefined
 		private destroyed: boolean = false
 
 		constructor(props) {
@@ -530,12 +539,14 @@ export default translateWithTracker<ISystemStatusProps, ISystemStatusState, ISys
 
 			this.state = {
 				systemStatus: undefined,
+				deviceDebugState: new Map(),
 			}
 		}
 
 		componentDidMount() {
 			this.refreshSystemStatus()
 			this.refreshInterval = setInterval(this.refreshSystemStatus, 5000)
+			this.refreshDebugStatesInterval = setInterval(this.refreshDebugStates, 1000)
 
 			// Subscribe to data:
 			this.subscribe(PubSub.peripheralDevices, {})
@@ -543,6 +554,7 @@ export default translateWithTracker<ISystemStatusProps, ISystemStatusState, ISys
 
 		componentWillUnmount() {
 			if (this.refreshInterval) clearInterval(this.refreshInterval)
+			if (this.refreshDebugStatesInterval) clearInterval(this.refreshDebugStatesInterval)
 			this.destroyed = true
 		}
 
@@ -570,6 +582,25 @@ export default translateWithTracker<ISystemStatusProps, ISystemStatusState, ISys
 					)
 					return
 				})
+		}
+
+		refreshDebugStates = () => {
+			for (const device of this.props.devices) {
+				if (device.type === PeripheralDeviceType.PLAYOUT && device.settings && device.settings['debugState']) {
+					MeteorCall.systemStatus
+						.getDebugStates(device._id)
+						.then((res) => {
+							const states: Map<PeripheralDeviceId, object> = new Map()
+							for (const [key, state] of Object.entries(res)) {
+								states.set(protectString(key), state)
+							}
+							this.setState({
+								deviceDebugState: states,
+							})
+						})
+						.catch((err) => console.log(`Error fetching device states: ${err}`))
+				}
+			}
 		}
 
 		renderPeripheralDevices() {
@@ -607,6 +638,7 @@ export default translateWithTracker<ISystemStatusProps, ISystemStatusState, ISys
 						device={d.device}
 						toplevel={toplevel}
 						hasChildren={d.children.length !== 0}
+						debugState={this.state.deviceDebugState.get(d.device._id)}
 					/>,
 				]
 				if (d.children.length) {