+++ meteor/server/systemStatus/systemStatus.ts
@@ -28,8 +28,11 @@ import { resolveCredentials, Credentials } from '../security/lib/credentials'
 import { SystemReadAccess } from '../security/system'
 import { StatusCode } from '@sofie-automation/blueprints-integration'
 import { PeripheralDevices, Workers, WorkerThreadStatuses } from '../collections'
-import { StudioId } from '@sofie-automation/corelib/dist/dataModel/Ids'
 import { getUpgradeSystemStatusMessages } from '../migration/upgrades'
+import { PeripheralDeviceId, StudioId } from '@sofie-automation/corelib/dist/dataModel/Ids'
+import { ServerPeripheralDeviceAPI } from '../api/peripheralDevice'
+import { PeripheralDeviceContentWriteAccess } from '../security/peripheralDevice'
+import { MethodContext } from '../../lib/api/methods'
 
 const PackageInfo = require('../../package.json')
 const integrationVersionRange = parseCoreIntegrationCompatabilityRange(PackageInfo.version)
@@ -398,4 +401,11 @@ export function status2ExternalStatus(statusCode: StatusCode): ExternalStatus {
 		return 'FAIL'
 	}
 	return 'UNDEFINED'
+}
+export async function getDebugStates(
+	methodContext: MethodContext,
+	peripheralDeviceId: PeripheralDeviceId
+): Promise<object> {
+	const access = await PeripheralDeviceContentWriteAccess.peripheralDevice(methodContext, peripheralDeviceId)
+	return ServerPeripheralDeviceAPI.getDebugStates(access)
 }